Сделана связь один ко многим по тз, не работает мод.окно,
