+++ FurnitureMovement/Data/OrderContext.cs
@@ -4,14 +4,13 @@ namespace FurnitureMovement.Data;
 
 public class OrderContext : DbContext
 {
-    // DbSet-свойства добавляем после описания наших табличных классов, сейчас не нужно
-    public DbSet<Order> Orders { get; set; }
-
-    // Обязательная настройка конструктора. Конфигуратор в Startup’е использует этот конструктор
     public OrderContext(DbContextOptions<OrderContext> options) : base(options)
     {
 
     }
+
+    public DbSet<Order> Orders { get; set; }
+    public DbSet<OrderFurniture> OrderFurnitures { get; set; }  
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
